Rename `monet.hpp` to `monet.h`
