+++ CMakeLists.txt
@@ -9,7 +9,7 @@ set(ZIOTOM78_MONET_INCLUDE_BUILD_DIR "${PROJECT_SOURCE_DIR}/include/")
 
 add_library(${ZIOTOM78_MONET_TARGET_NAME} INTERFACE)
 target_sources(${ZIOTOM78_MONET_TARGET_NAME}
-  INTERFACE "${PROJECT_SOURCE_DIR}/include/monet.hpp")
+  INTERFACE "${PROJECT_SOURCE_DIR}/include/monet.h")
 target_include_directories(${ZIOTOM78_MONET_TARGET_NAME}
   INTERFACE ${ZIOTOM78_MONET_INCLUDE_BUILD_DIR})
 target_compile_features(${ZIOTOM78_MONET_TARGET_NAME} INTERFACE cxx_std_11)
