+++ docs/index.adoc
@@ -24,7 +24,7 @@ You can proceed with the link:./tutorial.md[tutorial], or you can
 directly download the library and start coding:
 
 ....
-curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.hpp > monet.hpp
+curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.h > monet.h
 ....
 include::tutorial.adoc[]
