+++ docs/index.html
@@ -572,7 +572,7 @@ directly download the library and start coding:</p>
 </div>
 <div class="literalblock">
 <div class="content">
-<pre>curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.hpp &gt; monet.hpp</pre>
+<pre>curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.h &gt; monet.h</pre>
 </div>
 </div>
 </div>
@@ -592,7 +592,7 @@ Monet header file in it with the following command:</p>
 </div>
 <div class="literalblock">
 <div class="content">
-<pre>curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.hpp &gt; monet.hpp</pre>
+<pre>curl https://raw.githubusercontent.com/ziotom78/monet/master/include/monet.h &gt; monet.h</pre>
 </div>
 </div>
 <div class="paragraph">
@@ -602,7 +602,7 @@ favourite editor to create the following C++ file, and name it
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 
 <span class="directive">using</span> <span class="keyword">namespace</span> monet;
 
@@ -741,7 +741,7 @@ to show the coordinates of a few points in a canvas.</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;cstdio&gt;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;vector&gt;</span>
 
@@ -902,7 +902,7 @@ the saturation:</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;cstdio&gt;</span>
 
 <span class="directive">using</span> <span class="keyword">namespace</span> monet;
@@ -1003,7 +1003,7 @@ example:</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;cstdio&gt;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;vector&gt;</span>
 
@@ -1136,7 +1136,7 @@ you must provide two points: the <code>target'' point and the </code>end'' point
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;cstdio&gt;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;vector&gt;</span>
 
@@ -1216,7 +1216,7 @@ the <code>settransparency</code> function.</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 
 <span class="directive">using</span> <span class="keyword">namespace</span> monet;
 
@@ -1494,7 +1494,7 @@ the following example:</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;iostream&gt;</span>
 
 <span class="directive">using</span> <span class="keyword">namespace</span> monet;
@@ -1670,7 +1670,7 @@ following fields:</p>
 </div>
 <div class="listingblock">
 <div class="content">
-<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.hpp&quot;</span>
+<pre class="CodeRay highlight"><code data-lang="c++"><span class="preprocessor">#include</span> <span class="include">&quot;monet.h&quot;</span>
 <span class="preprocessor">#include</span> <span class="include">&lt;iostream&gt;</span>
 
 <span class="directive">using</span> <span class="keyword">namespace</span> monet;
