+++ docs/ref-color-table.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <cstdio>
 #include <string>
 #include <vector>
