+++ docs/ref-point.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <iostream>
 
 using namespace monet;
