+++ docs/tut02-bad.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <cstdio>
 #include <vector>
