+++ docs/tut03-no-saturation.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <cstdio>
 
 using namespace monet;
