+++ docs/tut04-closepath.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <cstdio>
 #include <vector>
