+++ docs/tut04-fillandstroke.cpp
@@ -1,4 +1,4 @@
-#include "monet.hpp"
+#include "monet.h"
 #include <cstdio>
 #include <vector>
