+++ examples/demo.cpp
@@ -1,4 +1,4 @@
-#include <monet.hpp>
+#include <monet.h>
 
 using namespace monet;
