+++ test/src/test-transforms.cpp
@@ -1,5 +1,5 @@
 #include <cassert>
-#include <monet.hpp>
+#include <monet.h>
 
 using namespace monet;
